+++ frontend/clockko-wellness-app/src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './typeReward'
+export * from './typeLandingPage'
+export * from './typesGlobal'
