+++ frontend/clockko-wellness-app/src/types/typeLandingPage.ts
@@ -1,0 +1,21 @@
+
+// Landing Page
+export interface LandingPageContent {
+  features: LandingFeature[];
+  heroText: string;
+  callToActionText: string;
+  testimonials?: Testimonial[];
+}
+
+export interface LandingFeature {
+  title: string;
+  description: string;
+  iconUrl?: string;
+}
+
+export interface Testimonial {
+  id: string;
+  author: string;
+  quote: string;
+  avatarUrl?: string;
+}
