+++ frontend/clockko-wellness-app/src/types/typeReward.ts
@@ -1,0 +1,11 @@
+// Reward System
+export interface Reward {
+  id: string;
+  userId: string;
+  points: number;
+  redeemed: RedeemedPerk[];
+}
+export interface RedeemedPerk {
+  perk: string;
+  date: Date;
+}
